Update Django W7

+++ page/views.py
@@ -1,5 +1,7 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Posting
+# [코드 추가] forms.py의 PostingForm 불러오기
+from .forms import PostingForm
 
 # Create your views here.
 def index(request):
@@ -8,16 +10,25 @@ def index(request):
 # Create
 def posting_create(request):
     if request.method == 'POST':
-        posting_title = request.POST.get('title')
-        posting_content = request.POST.get('content')
+        # [코드 작성] POST 방식으로 넘어온 값을 form에 저장
+        form = PostingForm(request.POST)
 
-        Posting.objects.create(
-            title=posting_title,
-            content=posting_content,
-        )
-        return redirect('page:posting_list')
-    else:
-        return render(request, 'page/posting_create.html')
+        # [코드 작성] form이 유효한지 확인
+        if form.is_valid():
+            # [코드 작성] posting 객체 생성
+            posting = Posting()
+            posting.title = form.cleaned_data.get('title')
+            posting.content = form.cleaned_data.get('content')
+
+            # [코드 작성] posting 객체 저장
+            posting.save()
+            return redirect('page:posting_list')
+    
+    form = PostingForm()
+    context = {
+        'form': form,
+    }
+    return render(request, 'page/posting_form.html', context)
 
 # Read
 def posting_list(request):
@@ -29,7 +40,7 @@ def posting_list(request):
 
 # Read
 def posting_detail(request, posting_id):
-    posting = Posting.objects.get(id=posting_id)
+    posting = get_object_or_404(Posting, id=posting_id)
     context = {
         'posting': posting,
     }
@@ -37,24 +48,28 @@ def posting_detail(request, posting_id):
 
 # Update
 def posting_update(request, posting_id):
-    posting = Posting.objects.get(id=posting_id)
+    posting = get_object_or_404(Posting, id=posting_id)
 
     if request.method == 'POST':
-        posting_title = request.POST.get('title')
-        posting_content = request.POST.get('content')
+        # [코드 작성] POST 방식으로 넘어온 값을 form에 저장
+        form = PostingForm(request.POST, instance=posting)
 
-        posting.title = posting_title
-        posting.content = posting_content
-        posting.save()
-        return redirect('page:posting_detail', posting_id)
+        # [코드 작성] form이 유효한지 확인
+        if form.is_valid():
+            form.save()
+            return redirect('page:posting_detail', posting_id)
     else:
-        context = {
-            'posting': posting,
-        }
-        return render(request, 'page/posting_update.html', context)
+        form = PostingForm(instance=posting)
+
+    context = {
+        'form': form,
+    }
+    return render(request, 'page/posting_form.html', context)
 
 # Delete
 def posting_delete(request, posting_id):
-    posting = Posting.objects.get(id=posting_id)
-    posting.delete()
-    return redirect('page:posting_read')
+    if request.method == 'POST':
+        posting = get_object_or_404(Posting, id=posting_id)
+        posting.delete()
+        return redirect('page:posting_list')
+    return redirect('page:posting_detail', posting_id)
